derive the agent population from market cap when the count is zero
An agent is a unit of resolution, not a person: share float times start price
says how big the market is, agent count says how finely it is resolved, and
every per-agent quantity follows from the two. Deriving one from the other is
now offered rather than imposed, since only a narrow band of market caps can
support it -- below about a million the market is too thin to be worth
simulating, and a two billion dollar cap implies enough agents to make a run
take minutes per simulated day.

An agent count of 0 sizes the population to the market, at the point where the
cash bands read as literal dollars for the expected type mix, so the derived
market is self consistent by construction rather than needing a new constant.
Clamped at 25 and 50,000, the latter being roughly forty seconds of back data
per day. Deliberately a function of price and float only and never of the
seed, so the reset dialog can show the number a run will use before that run
has drawn anything.

agentStartCount widens from unsigned short to unsigned int throughout, so
counts above 65,535 stop wrapping. Verified with a 70,000 agent run, which the
old type would have turned into 4,464.

The reset dialog shows the derived count, an estimated back data time, and
whether the value was floored or capped, since agent count is the field that
decides how long a run takes and the only one whose value was invisible until
after pressing enter.

Also documents the resolution framing in the readme, where it explains what
the two parameters mean to each other.

+++ Core/include/CoreSim.h
@@ -27,7 +27,7 @@ struct Config {
     Session liveStartSession = Session::REGULAR;
     /* Minimum resting orders per side required at handoff, 0 disables the check */
     unsigned int minLiquidity = 0;
-    unsigned short agentStartCount = 100;
+    unsigned int agentStartCount = 100;
     unsigned int obShareFloat = 100'000;
     double obStartPrice = 1.00;
     /* MEDIAN share of the resident population that is transient, 0 disables
@@ -159,7 +159,7 @@ public:
 
     // ---- Simulation Initialization Functions ----
 
-    void initAgents(unsigned short _agentStartCount);
+    void initAgents(unsigned int _agentStartCount);
     /* Give an agent its apportioned starting position, with a plausible cost basis
     *
     * Split out of initAgents because the float can only be apportioned once every weight
@@ -322,7 +322,7 @@ public:
         unsigned int backDataDays = 1,
         Session liveStartSession = Session::REGULAR,
         unsigned int minLiquidity = 0,
-        unsigned short agentStartCount = 100,
+        unsigned int agentStartCount = 100,
         unsigned int obShareFloat = 250'000,
         double obStartPrice = 1.00,
         double transientFraction = 0.0
